+++ app/src/main/java/io/github/frankolt/githubexplorer/data/http/github/models/SearchResultItemResponse.kt
@@ -5,7 +5,7 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class SearchResultItemResponse(
-    @Json(name = "id") val id: Int? = null,
+    @Json(name = "id") val id: Long? = null,
     @Json(name = "node_id") val nodeId: String? = null,
     @Json(name = "name") val name: String? = null,
     @Json(name = "full_name") val fullName: String? = null,
@@ -19,13 +19,13 @@ data class SearchResultItemResponse(
     @Json(name = "updated_at") val updatedAt: String? = null,
     @Json(name = "pushed_at") val pushedAt: String? = null,
     @Json(name = "homepage") val homepage: String? = null,
-    @Json(name = "size") val size: Int? = null,
-    @Json(name = "stargazers_count") val stargazersCount: Int? = null,
-    @Json(name = "watchers_count") val watchersCount: Int? = null,
+    @Json(name = "size") val size: Long? = null,
+    @Json(name = "stargazers_count") val stargazersCount: Long? = null,
+    @Json(name = "watchers_count") val watchersCount: Long? = null,
     @Json(name = "language") val language: String? = null,
-    @Json(name = "forks_count") val forksCount: Int? = null,
-    @Json(name = "open_issues_count") val openIssuesCount: Int? = null,
+    @Json(name = "forks_count") val forksCount: Long? = null,
+    @Json(name = "open_issues_count") val openIssuesCount: Long? = null,
     @Json(name = "master_branch") val masterBranch: String? = null,
     @Json(name = "default_branch") val defaultBranch: String? = null,
-    @Json(name = "score") val score: Int? = null
+    @Json(name = "score") val score: Long? = null
 )
